+++ JustLab/Math/Matrix.h
@@ -8,19 +8,19 @@
 #include "Vector.h"
 
 namespace Just {
-    template<size_t ROW, size_t COL, typename T>
+    template<uint32_t ROW, uint32_t COL, typename T>
     struct Matrix {
         T data[ROW][COL];
 
         Matrix() {
-            for (size_t row = 0; row < ROW; row++)
-                for (size_t col = 0; col < COL; col++)
+            for (uint32_t row = 0; row < ROW; row++)
+                for (uint32_t col = 0; col < COL; col++)
                     data[row][col] = 0;
         }
 
         Matrix(const std::initializer_list<Vector<COL, T>> &mat) {
             auto it = mat.begin();
-            for (size_t row = 0; row < ROW; row++)
+            for (uint32_t row = 0; row < ROW; row++)
                 SetRow(row, *it++);
         }
 
@@ -28,41 +28,41 @@ namespace Just {
 
         Matrix &operator=(const Matrix<ROW, COL, T> &other) = default;
 
-        T *operator[](size_t row) {
+        T *operator[](uint32_t row) {
             assert(row < ROW);
             return data[row];
         }
 
-        const T *operator[](size_t row) const {
+        const T *operator[](uint32_t row) const {
             assert(row < ROW);
             return data[row];
         }
 
-        Vector<COL, T> Row(size_t row) const {
+        Vector<COL, T> Row(uint32_t row) const {
             assert(row < ROW);
             Vector<COL, T> tmp;
-            for (size_t col = 0; col < COL; col++)
+            for (uint32_t col = 0; col < COL; col++)
                 tmp[col] = data[row][col];
             return tmp;
         }
 
-        Vector<ROW, T> Col(size_t col) const {
+        Vector<ROW, T> Col(uint32_t col) const {
             assert(col < COL);
             Vector<ROW, T> tmp;
-            for (size_t row = 0; row < ROW; row++)
+            for (uint32_t row = 0; row < ROW; row++)
                 tmp[row] = data[row][col];
             return tmp;
         }
 
-        void SetRow(size_t row, const Vector<COL, T> &v) {
+        void SetRow(uint32_t row, const Vector<COL, T> &v) {
             assert(row < ROW);
-            for (size_t col = 0; col < COL; col++)
+            for (uint32_t col = 0; col < COL; col++)
                 data[row][col] = v[col];
         }
 
-        void SetCol(size_t col, const Vector<ROW, T> &v) {
+        void SetCol(uint32_t col, const Vector<ROW, T> &v) {
             assert(col < COL);
-            for (size_t row = 0; row < ROW; row++)
+            for (uint32_t row = 0; row < ROW; row++)
                 data[row][col] = v[row];
         }
     };
