+++ JustLab/Math/VectorExt.h
@@ -8,10 +8,10 @@
 
 namespace Just {
     //输出
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline std::ostream &operator<<(std::ostream &os, const Vector<N, T> &v) {
         os << "(";
-        for (size_t i = 0; i < N; i++) {
+        for (uint32_t i = 0; i < N; i++) {
             os << v.data[i];
             if (i + 1 != N)
                 os << ",";
@@ -21,153 +21,153 @@ namespace Just {
     }
 
     //取反
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> operator-(const Vector<N, T> &v) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = -v[i];
         return tmp;
     }
 
     //加法
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> operator+(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] + v2[i];
         return tmp;
     }
 
     //减法
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> operator-(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] - v2[i];
         return tmp;
     }
 
     //乘法
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> operator*(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] * v2[i];
         return tmp;
     }
 
     //赋值加
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> &operator+=(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             v1[i] += v2[i];
         return v1;
     }
 
     //赋值减
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> &operator-=(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++) {
+        for (uint32_t i = 0; i < N; i++) {
             v1[i] -= v2[i];
         }
         return v1;
     }
 
     //赋值乘
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> &operator*=(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             v1[i] *= v2[i];
         return v1;
     }
 
     //数乘
-    template<size_t N, typename T, typename U>
+    template<uint32_t N, typename T, typename U>
     inline Vector<N, T> operator*(const Vector<N, T> &v1, U k) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] * static_cast<T>(k);
         return tmp;
     }
 
-    template<size_t N, typename T, typename U>
+    template<uint32_t N, typename T, typename U>
     inline Vector<N, T> operator*(U k, const Vector<N, T> &v1) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] * static_cast<T>(k);
         return tmp;
     }
 
-    template<size_t N, typename T, typename U>
+    template<uint32_t N, typename T, typename U>
     inline Vector<N, T> operator/(const Vector<N, T> &v1, U k) {
         Vector<N, T> tmp;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             tmp[i] = v1[i] / static_cast<T>(k);
         return tmp;
     }
 
     //赋值数乘
-    template<size_t N, typename T, typename U>
+    template<uint32_t N, typename T, typename U>
     inline Vector<N, T> &operator*=(Vector<N, T> &v, U k) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             v[i] *= static_cast<T>(k);
         return v;
     }
 
-    template<size_t N, typename T, typename U>
+    template<uint32_t N, typename T, typename U>
     inline Vector<N, T> &operator/=(Vector<N, T> &v, U k) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             v[i] /= static_cast<T>(k);
         return v;
     }
 
     //比较
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline bool operator<(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             if (v1[i] >= v2[i])
                 return false;
         return true;
     }
 
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline bool operator>(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             if (v1[i] <= v2[i])
                 return false;
         return true;
     }
 
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline bool operator<=(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             if (v1[i] > v2[i])
                 return false;
         return true;
     }
 
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline bool operator>=(Vector<N, T> &v1, const Vector<N, T> &v2) {
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             if (v1[i] < v2[i])
                 return false;
         return true;
     }
 
     //长度
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline T Length(const Vector<N, T> &v) {
         T sum = 0;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             sum += v[i] * v[i];
         return std::sqrt(sum);
     }
 
     //点积
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline T Dot(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         T sum = 0;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             sum += v1[i] * v2[i];
         return sum;
     }
@@ -185,28 +185,28 @@ namespace Just {
     }
 
     //归一化
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> Normalize(const Vector<N, T> &v) {
         T sum = 0;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             sum += v[i] * v[i];
         return v / std::sqrt(sum);
     }
 
     //各维度最小值
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> MinValue(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         Vector<N, T> min;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             min[i] = std::min(v1[i], v2[i]);
         return min;
     }
 
     //各维度最大值
-    template<size_t N, typename T>
+    template<uint32_t N, typename T>
     inline Vector<N, T> MaxValue(const Vector<N, T> &v1, const Vector<N, T> &v2) {
         Vector<N, T> max;
-        for (size_t i = 0; i < N; i++)
+        for (uint32_t i = 0; i < N; i++)
             max[i] = std::max(v1[i], v2[i]);
         return max;
     }
